+++ app/builders/messages/facebook/message_builder.rb
@@ -23,6 +23,7 @@ class Messages::Facebook::MessageBuilder < Messages::Messenger::MessageBuilder
 
     ActiveRecord::Base.transaction do
       build_contact_inbox
+      refresh_contact_name_if_needed
       build_message
     end
   rescue Koala::Facebook::AuthenticationError => e
@@ -120,37 +121,50 @@ class Messages::Facebook::MessageBuilder < Messages::Messenger::MessageBuilder
 
   def process_contact_params_result(result)
     {
-      name: "#{result['first_name'] || 'John'} #{result['last_name'] || 'Doe'}",
+      name: Facebook::ContactNameResolver.resolve(result),
       account_id: @inbox.account_id,
       avatar_url: result['profile_pic']
     }
   end
 
+  def refresh_contact_name_if_needed
+    contact = @contact_inbox.contact
+    new_name = contact_params[:name]
+    return if new_name.blank? || new_name == contact.name
+    return unless Facebook::ContactNameResolver.refreshable_fallback?(contact.name)
+
+    contact.update!(name: new_name)
+  end
+
   # rubocop:disable Metrics/AbcSize
   # rubocop:disable Metrics/MethodLength
   def contact_params
-    begin
-      k = Koala::Facebook::API.new(@inbox.channel.page_access_token) if @inbox.facebook?
-      result = k.get_object(@sender_id) || {}
-    rescue Koala::Facebook::AuthenticationError => e
-      Rails.logger.warn("Facebook authentication error for inbox: #{@inbox.id} with error: #{e.message}")
-      Rails.logger.error e
-      @inbox.channel.authorization_error!
-      raise
-    rescue Koala::Facebook::ClientError => e
-      result = {}
-      # OAuthException, code: 100, error_subcode: 2018218, message: (#100) No profile available for this user
-      # We don't need to capture this error as we don't care about contact params in case of echo messages
-      if e.message.include?('2018218')
-        Rails.logger.warn e
-      else
-        ChatwootExceptionTracker.new(e, account: @inbox.account).capture_exception unless @outgoing_echo
-      end
-    rescue StandardError => e
-      result = {}
-      ChatwootExceptionTracker.new(e, account: @inbox.account).capture_exception
+    @contact_params ||= begin
+      result = fetch_facebook_profile_result
+      process_contact_params_result(result)
     end
-    process_contact_params_result(result)
+  end
+
+  def fetch_facebook_profile_result
+    k = Koala::Facebook::API.new(@inbox.channel.page_access_token) if @inbox.facebook?
+    k.get_object(@sender_id) || {}
+  rescue Koala::Facebook::AuthenticationError => e
+    Rails.logger.warn("Facebook authentication error for inbox: #{@inbox.id} with error: #{e.message}")
+    Rails.logger.error e
+    @inbox.channel.authorization_error!
+    raise
+  rescue Koala::Facebook::ClientError => e
+    # OAuthException, code: 100, error_subcode: 2018218, message: (#100) No profile available for this user
+    # We don't need to capture this error as we don't care about contact params in case of echo messages
+    if e.message.include?('2018218')
+      Rails.logger.warn e
+    else
+      ChatwootExceptionTracker.new(e, account: @inbox.account).capture_exception unless @outgoing_echo
+    end
+    {}
+  rescue StandardError => e
+    ChatwootExceptionTracker.new(e, account: @inbox.account).capture_exception
+    {}
   end
   # rubocop:enable Metrics/AbcSize
   # rubocop:enable Metrics/MethodLength