+++ lib/optimia.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module Optimia
+  VERSION = '0.1.10'
+  STAGING_IMAGE_TAG = 'staging-cw-4.10.1-optimia-0.1.10'
+end